[LETS-719] [LETS-720] Refactor to decouple [dis]connection and the main connection (#4390)
http://jira.cubrid.org/browse/LETS-719

http://jira.cubrid.org/browse/LETS-720

Note that this PR includes both inflating the connection_handler vector in advance and changing where to change the main connection. These both are for decoupling. I was going to split them into two PRs, but one required adding logic and code which are bound to be removed in the other and ensuring it works, so I've combined them. For LETS-720 in this PR, only push_request/send_receive are changed and the main connection change is removed when handling the disconnect msg.

In addition: The internal state of connection_handler is defined. A connection_handler must be in one of those states. See the comment for tran_server::connection_handler::state.