[LETS-754] Reset main connection after connecting PSes (#4493)
http://jira.cubrid.org/browse/LETS-754

Reset the main connection after connecting all PSes.

- Reset the main connection after establishing all connections on boot()
- Reset the main connection after checking and reconnecting all PSes by ps_connector.
- Fix reset_main_connection not to mix up the main connection by widening the lock scope. It was that a main connection elected first can be set later.

Sidenote:
When the catch-up process is implemented, the resetting should be also revised. We will not be able to set a new connection as the main connection until the asynchronous catch-up process is done. My plan for now is to extend ps_connector to also check and reset the main connection periodically.